fix: remove signup page in favor of single login/signup page
